requirements.txt
ImageHash==4.3.2
numpy==2.2.3
pillow==11.1.0
PyWavelets==1.8.0
scipy==1.15.2
